:recycle: api: Move add-certification-candidate-to-session injection to session
