+++ api/src/certification/session/infrastructure/repositories/certification-cpf-city-repository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
-import { CertificationCpfCity } from '../../domain/models/CertificationCpfCity.js';
+import { CertificationCpfCity } from '../../../shared/domain/models/CertificationCpfCity.js';
 
 const COLUMNS = ['id', 'name', 'postalCode', 'INSEECode', 'isActualName'];
