+++ api/src/certification/session/infrastructure/repositories/certification-cpf-country-repository.js
@@ -1,5 +1,5 @@
 import { knex } from '../../../../../db/knex-database-connection.js';
-import { CertificationCpfCountry } from '../../../../certification/shared/domain/models/CertificationCpfCountry.js';
+import { CertificationCpfCountry } from '../../../shared/domain/models/CertificationCpfCountry.js';
 
 const getByMatcher = async function ({ matcher }) {
   const COLUMNS = ['id', 'code', 'commonName', 'originalName', 'matcher'];
