+++ api/tests/certification/session/unit/domain/usecases/validate-sessions_test.js
@@ -3,8 +3,8 @@ import { CERTIFICATION_SESSIONS_ERRORS } from '../../../../../../lib/domain/cons
 import { CertificationCandidate } from '../../../../../../lib/domain/models/CertificationCandidate.js';
 import { Session } from '../../../../../../src/certification/session/domain/models/Session.js';
 import { SessionMassImportReport } from '../../../../../../src/certification/session/domain/models/SessionMassImportReport.js';
+import { CpfBirthInformationValidation } from '../../../../../../src/certification/session/domain/services/certification-cpf-service.js';
 import { validateSessions } from '../../../../../../src/certification/session/domain/usecases/validate-sessions.js';
-import { CpfBirthInformationValidation } from '../../../../../../src/certification/shared/domain/services/certification-cpf-service.js';
 import { domainBuilder, expect, sinon } from '../../../../../test-helper.js';
 import { getI18n } from '../../../../../tooling/i18n/i18n.js';
