+++ api/tests/certification/shared/integration/infrastructure/repositories/certification-cpf-city-repository_test.js
@@ -1,5 +1,5 @@
+import * as certificationCpfCityRepository from '../../../../../../src/certification/session/infrastructure/repositories/certification-cpf-city-repository.js';
 import { CertificationCpfCity } from '../../../../../../src/certification/shared/domain/models/CertificationCpfCity.js';
-import * as certificationCpfCityRepository from '../../../../../../src/certification/shared/infrastructure/repositories/certification-cpf-city-repository.js';
 import { databaseBuilder, domainBuilder, expect } from '../../../../../test-helper.js';
 
 describe('Integration | Repository | certificationCpfCityRepository', function () {
