+++ api/tests/unit/domain/usecases/correct-candidate-identity-in-certification-course_test.js
@@ -1,7 +1,7 @@
 import { CertificationCandidatesError } from '../../../../lib/domain/errors.js';
 import { CertificationCourse } from '../../../../lib/domain/models/CertificationCourse.js';
 import { correctCandidateIdentityInCertificationCourse } from '../../../../lib/domain/usecases/correct-candidate-identity-in-certification-course.js';
-import { CpfBirthInformationValidation } from '../../../../src/certification/shared/domain/services/certification-cpf-service.js';
+import { CpfBirthInformationValidation } from '../../../../src/certification/session/domain/services/certification-cpf-service.js';
 import { catchErr, domainBuilder, expect, sinon } from '../../../test-helper.js';
 
 describe('Unit | UseCase | correct-candidate-identity-in-certification-course', function () {
